Can use. Validate tesselate
