+++ sources/include/citygml/citygml.h
@@ -69,6 +69,7 @@ namespace citygml
             , minLOD( 0 )
             , maxLOD( 4 )
             , optimize( false )
+            , tesselate( false )
             , pruneEmptyObjects( false )
             , destSRS( "" )
             , srcSRS( "" )
