+++ sources/src/parser/citygmldocumentparser.cpp
@@ -126,7 +126,7 @@ namespace citygml {
             tesselator.setKeepVertices(m_parserParams.keepVertices);
 
             CITYGML_LOG_INFO(m_logger, "Start postprocessing of the citymodel.");
-            m_rootModel->finish(tesselator, m_parserParams.optimize, m_logger);
+            m_rootModel->finish(tesselator, m_parserParams.optimize, m_parserParams.tesselate, m_logger);
             CITYGML_LOG_INFO(m_logger, "Finished postprocessing of the citymodel.");
 
             m_rootModel->setThemes(m_factory->getAllThemes());
